requirements.txt
# Customer Churn Prediction - Python Dependencies

# Core Data Science Libraries
pandas>=1.3.0
numpy>=1.21.0

# Machine Learning
scikit-learn>=1.0.0
xgboost>=1.5.0
imbalanced-learn>=0.9.0

# Visualization
matplotlib>=3.4.0
seaborn>=0.11.0

# Model Interpretability (optional but recommended)
shap>=0.40.0

# Deployment
streamlit>=1.10.0

# Utilities
joblib>=1.1.0
python-dotenv>=0.19.0
toml>=0.10.2

# Jupyter (for notebooks)
jupyter>=1.0.0
notebook>=6.4.0
ipykernel>=6.0.0

# Additional utilities
scipy>=1.7.0
tqdm>=4.62.0
